Read the Postgres error log rather than HTTP status: most GETs answer
200 {status:"setup_required"}, which hid broken SQL behind a healthy code.

- users RLS meant a team could not see its own people. 221 'JOIN users' sites
  silently dropped every row but your own, so rosters, assignments and
  attribution came back empty. 0517 adds a co-membership SELECT policy through
  a SECURITY DEFINER helper with a pinned search_path.
- private pEPA was computed and then discarded every request: scout_sample is
  integer and the value is fractional, and the failure was swallowed by a bare
  catch, which also aborted later writes in the same transaction (0518).
- org creation always failed and invited scouts could never join (0515, 0516).
- Replaying all 362 migrations from zero also surfaced two historical breaks:
  0113 used a reserved word unquoted and 0158 changed a view's column order
  under CREATE OR REPLACE. Both are inert for an already-migrated database and
  are the difference between a fresh environment building or not.

+++ apps/web/app/api/org/analytics/private-epa/route.ts
@@ -46,7 +46,7 @@ export async function GET(request: Request) {
           computedAt: string;
         }>(
           `SELECT team_key AS "teamKey", public_epa::float AS "publicEpa", pepa::float AS pepa,
-                  scout_component_epa::float AS "scoutComponentEpa", scout_sample AS "scoutSample",
+                  scout_component_epa::float AS "scoutComponentEpa", scout_sample::float AS "scoutSample",
                   components, computed_at::text AS "computedAt"
            FROM private_epa_snapshots
            WHERE org_id = $1 AND event_key = $2
